jacobi version with int() and various scripts for plots

+++ jacobi.h
@@ -1,6 +1,6 @@
 #ifndef __JACOBI_H
 #define __JACOBI_H
 
-double jacobi(int N, double **array_in, double **array_out, int max_iter, double threshold, double **fmatrix);
+int jacobi(int N, double **array_in, double **array_out, int max_iter, double threshold, double **fmatrix);
 
 #endif
